+++ CHANGES.rst
@@ -4,9 +4,21 @@ Changes
 4.2.0 (unreleased)
 ------------------
 
+- Raise more informative KeyError subclasses from the utility when intids
+  or objects cannot be found. This distinguishes them from errors
+  raised by normal dictionaries or BTrees, and is useful in unit
+  testing or when persisting intids or sharing them among processes
+  for later or concurrent use.
+
+- Propagate ``POSKeyError`` from ``queryId`` instead of returning the
+  default object. This exception indicates a corrupt database, not a
+  missing object. The ``queryObject`` function already behaved this way.
+
 - Add support for Python 3.5.
 
 - Drop support for Python 2.6.
+
+- Add support for PyPy3 5.2.
 
 - Stop depending on ZODB for anything except testing.